Demo_TurboVision/requirements.txt
azure-identity
azure-keyvault-secrets
azure-core
requests
pillow
matplotlib
moviepy
typing
opencv-python-headless
python-dotenv